+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `location` attributes.
+- `searchState` parsing.
+
 ## [1.5.1] - 2020-06-17
 
 ### Removed